backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.12.0
Pillow>=11.0.0
numpy>=2.1.0
scikit-learn>=1.4.0
tensorflow>=2.17.0
tink>=1.7.0
cryptography>=41.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
whitenoise>=6.6.0
PyJWT>=2.8.0
